Remove current-file from SEL project objects
This removes some potentially confusing behavior in which projects
would override `no-applicable-method` to dispatch to the
current-file (if set).  This makes it hard to have software objects
inherit from both a project and from the type of that project.

This did require the addition of two explicit function definitions for
clang-projects.  Specifically, `update-asts` and `update-caches`.