Minor fix to multithreaded instrumentation
Instrument the files in the project which do not contain
a program entry point (main function) prior to instrumenting
those files which do.  This ensures that the names and types
given at the beginning of the trace contain all of the
information needed for trace collection.

Previously, the trace did not contain all of the necessary
information as the names and types were written to file
prior to being completely populated.