Adding initial Javascript software object
This commit adds initial support for JavaScript software
objects to the software evolution library.  This initial
implementation supports single-file JavaScript software
objects along with instrumentation and trace collection
for JavaScript.  Clients should consider all methods and
functionality to be in the prototype phase.

The javascript.lisp file implements the core functionality
of the software-evolution-library required by Bug-Injector
for files written in JavaScript.  The core functionality
is supported by the command line tool acorn
(https://github.com/acornjs/acorn.git), which parses
the code into ASTs.

As with clang software objects, javascript software
objects are a parsed, AST representation.  All of the generic
methods for parseable software objects required by Bug-Injector
have been implemented (in prototype form), allowing for a rich
variety of manipulations and tests.

Every AST in a javascript software object contains an AST class
field with the AST type (e.g. BlockStatement, ForStatement).
Additional fields given by the acorn parser can be found in an
association list on the AST's aux-data field.  For more information
on the ASTs, see
https://media.readthedocs.org/pdf/esprima/4.0/esprima.pdf.

Note: certain methods in this class are over-loaded to provide
consistency with other software objects, but are not used in this
implementation.

Instrumentation and trace collection utilize the s-expression
trace format, as with Java software objects.  Because this
is a prototype, the s-expression trace database is appealing
as it is quicker to standup.  Additionally, this ellides
any potential issues with string handling in the binary
trace database (untested functionality).  Finally, because
JavaScript is a dynamic language, we cannot know prior to
executing the program the names of all of the variables
and the fields on these variables.  Unfortunately, the binary
trace database requires this information at the beginning of
the trace, which may disallow us from utilizing the binary
trace format.  To support both trace formats, both
binary-traceable and sexp-traceable classes where added
for each type of trace.

Where possible, this commit rolls up generic functions
and their implementation to the abstract paraseable software
object and instrumentation interface to allow for sharing
between clang and javascript implementation.

This commit also includes unit tests and documentation updates.