Keep a persistent clack server for all REST tests
This removes some overhead and some risk of errors due to ports still
being occupied from previous tests.  It also wraps some simple error
handling around creation of the clack server.

It results in tests that are less well isolated (they all use the same
server), but maybe this will increase the detective power of the test
suite as if test b fails because of web-server state from test a
presumably that is something we would like to know about.