Drop qlot and USER_QUICK_LISP
This simplification of SEL's CI infrastructure removes the dependency
on qlot and removes the use of the USER_QUICK_LISP environment
variable which is redundant with the QUICK_LISP environment variable.

Note that the .ci/qlfile.{external,grammatech} files are still used,
however their contents are simply checked out directly into the
quicklisp/local-projects directory instead of using qlot.

The support in cl.mk for parsing/installing dependencies from qlfiles
now supports an optional fourth column which (if present) should name
a branch (default is "master").