Added features to Create Software API
Create Software REST API now supports CLANG-PROJECT.
It supports derived classes as well, such as bug-injector
defined BI-CLANG-TRACEABLE-STYLEABLE, and
BI-CLANG-TRACEABLE-PROJECT.
In the body of the API call (POST) you define compilation-database,
artifacts, build command, and clang-class.

All software objects get FORMAT-GENOME automatically applied when they
are created.
Strings which contain "::" are considered to be symbols
to be converted to Lisp in the designated package. This allows
software objects defined outside of SEL (by derived packages) to
be instantiated.