Adding a `parseable-project` class
This class will contain methods common to all projects which
may be parsed into ASTs.  For now, it simply contains the
`asts` method.

This commit also moves the `to-file` specialization from
`clang-project` to `project` software objects.  This specialization
should be at the `project` level.  Additionally, this commit performs
some refactoring to remove the need to set the *build-dir* variable
within this method.