Add several cl-fad functions to utility.lisp
These include `directory-wildcard` (a helper),
`list-directory`, and `walk-directory`.  Of these,
`walk-directory` is most general as it leverages
`list-directory` and, in turn, `directory-wildcard`.
`walk-directory` will be utilized to better identify
JavaScript files in a project.

These functions conflict with similarly defined functions
within osicat.  However, the osicat version of `walk-directory`
is insufficient for our purposes, as the full file path is not
passed to the function argument, making it impossible to know
where to find the file.

TMP