Improve identification of JS project files
We assume a file is in a JavaScript project if:
1. It is not in node_modules/ dependencies directory.
2. It is not in test/ or tests/ directory.
3. It ends in the js extension.
4. It is listed as a "bin" in package.json.
5. It is listed as "main" in package.json.

This commit also updates JavaScript project to creation
to use the `parsing-mode` field on JavaScript software objects.
For files which are not program entry points, this field is
set to :module; otherwise, Acorn will parse the file as a script.