Rolled up several commits for REST-API
When REST API calls fail, more informative information
will be captured. In addition, settings have been adjusted so
that, by default, the Lisp debugger will be invoked when a failure
occurs. This can be changed easily in a running server.

Introduced a new syntax <resource-name>:<oid> which can
be passed as a string field via the rest api. It will cause
a lookup in the client session for a resource of the
requested type, of the specified oid. In that case the
server will use the found resource.

Also fix tab/whitespace errors in rest.lisp.