Fixes to clang-project `collect-evolve-files`
This changeset makes a number of fixes to correctly identify files
in a clang-project using a compilation database, either user-specified
or created by watching the project compile with bear.

Most importantly, this restores the behavior of creating compilation
databases in a temporary sandbox.  We do not want to directly modify
the project on the filesystem unless explicitly directed to by
the client (e.g. via a `to-file` call).