Add extensions to API for ssr
Add &key &allow-other-keys for replace-ast, ast-diff,
other merge functions.
Add better control of string diffing.
Ability to turn off recontextualization in replace-ast,
but still using apply-mutation-op mechanism.
Add ast-ancestors, which gives a bottom up list of the
ancestors of an ast node.  This is used in ssr matching
that requires information on the context of the node
being matched.
Make replace-ast work if the node being replaced is the
root of the ast.