Remove duplicate keyword/dynamic-variable option
In the past we have run into problems when we have exposed multiple
ways to set an option like this.  I suggest we stick to just the
keyword option for now.  (I see that *diff-on-strings* is not exposed
by the package, but that doesn't necessarily mean someone won't set it
anyway.)