Add option for kill-process to kill process tree
The :children option, when true, causes it to kill
all descendants of the process, as well as the process
itself.  The test-suite code now uses this when
killing the test process, as previously a child might
have stayed alive, holding a stream, so reading
from that stream would never finish.