Fix race condition in `finish-test`
Previously, we would call kill-process function without the
:children keyword as a first attempt to kill the process.
This would kill the top-level process, but not the children.
Because the top-level process was killed, we would not call
kill-process a second time with the :children keyword present
and the SIGKILL signal to completely terminate the process.
Therefore, the process would hang attempting to read the
child process's stderr stream when using zlib with Bug-Injector.