Better error message when test script not found
Previously, we would get stack trace below, which is not
helpful.

The value NIL is not of type (OR (VECTOR CHARACTER) (VECTOR NIL) BASE-STRING PATHNAME STREAM) when binding SB-IMPL::PATHSPEC
Backtrace for: #<SB-THREAD:THREAD "main thread" RUNNING {10006305B3}>
0: (PROBE-FILE NIL) [external]
1: (SOFTWARE-EVOLUTION-LIBRARY/COMMAND-LINE:RESOLVE-TEST-SCRIPT-FROM-TEST-SCRIPT "gt-harnesssafd.sh test ~a ~d" #P"/home/jruchti/synth/benchmarks/grep/")
2: (BUG-INJECTOR/BI:BI #P"/home/jruchti/synth/benchmarks/grep/" (#<BUG-INJECTOR/SCION:CLANG-SCION BUG-INJECTOR/SCIONS/CLANG:ANTI-DEBUG-LINUX>) "gt-harnesssafd.sh test ~a ~d" "1" :DRY-RUN NIL :INSTRUMENT-GRAFTS NIL :JUST NIL :LIMIT 1 :STORE-TRACES NIL :LOAD-TRACES NIL :ONLY-FUNCS NIL :SKIP-FUNCS NIL :MERGE NIL :TRACE-STRINGS NIL :LANGUAGE "c" :HELP NIL :QUIET NIL :VERBOSE 2 :LOAD NIL :EVAL NIL :OUT-DIR (:ABSOLUTE "home" "jruchti" "synth" "quicklisp" "local-projects" "sel" "out") :READ-SEED NIL :SAVE-SEED NIL :INTERACTIVE NIL :MANUAL NIL :SWANK NIL :COMPILER "clang" :FLAGS NIL :BUILD-COMMAND "./gt-harness.sh build" :ARTIFACTS ("grep") :COMPILATION-DATABASE NIL :NUM-THREADS 1 :PROFILE NIL :NO-STORE-SOFTWARE NIL :SAVE-ORIGINAL T :MAX-TRACE-POINTS NIL)
3: (BUG-INJECTOR/BI:RUN-BI)
4: ((LAMBDA NIL :IN UIOP/IMAGE:RESTORE-IMAGE))
5: (UIOP/IMAGE:CALL-WITH-FATAL-CONDITION-HANDLER #<CLOSURE (LAMBDA NIL :IN UIOP/IMAGE:RESTORE-IMAGE) {1004155E8B}>)
6: ((FLET SB-UNIX::BODY :IN SB-EXT:SAVE-LISP-AND-DIE))
7: ((FLET "WITHOUT-INTERRUPTS-BODY-27" :IN SB-EXT:SAVE-LISP-AND-DIE))
8: ((LABELS SB-IMPL::RESTART-LISP :IN SB-EXT:SAVE-LISP-AND-DIE))
Above backtrace due to this condition:
The value NIL is not of type (OR (VECTOR CHARACTER) (VECTOR NIL) BASE-STRING PATHNAME STREAM) when binding SB-IMPL::PATHSPEC
The value NIL is not of type (OR (VECTOR CHARACTER) (VECTOR NIL) BASE-STRING PATHNAME STREAM) when binding SB-IMPL::PATHSPEC

In sel:c6209b9e, changes were made to add an additional `probe-file`
function call prior to asserting the test script exists, ostensibly
to handle symlinks.  Unfortunately, if the file does not exist,
`probe-file` will return nil, and the subsequent call to `probe-file`
as part of an assert will give the above backtrace which is unhelpful.

This commit restores the previous behavior which gives a helpful
error message, while preserving handling of symlinks.