Unescape the command field of compilation database
The commit ensures proper escaping of command line flags for use
with clang-mutate.

Per https://clang.llvm.org/docs/JSONCompilationDatabase.html,
the command field is "the compile command executed. After JSON
unescaping, this must be a valid command to rerun the exact
compilation step for the translation unit in the environment
the build system uses. Parameters use shell quoting and shell
escaping of quotes, with ‘"’ and ‘\’ being the only special
characters. Shell expansion is not supported."