Make Coq obj robust against nonreserializable ASTs
When loading from file, check if the AST will be reserializable (due to
bugs in SerAPI - filing issues will require that we upgrade to a newer
version first). If not, store the raw string in the genome, rather than
the AST so that we can properly write the genome to file.