Fix race in multi-threaded instrumentation
We *MUST* instrument the files in the order given by the
`instrumentation-files` method.  Previously, we would parse the files
in a multi-threaded manner; the results of parsing would be returned
in the order parsing completed instead of the order given by
`instrumentation-files`.  This led to issues later when using the
trace to index into the project.

We now perform parsing in an around method as a pre-step.