Don't set compiler/flags from compilation db
The compiler/flags fields are only present on `source` (i.e. single-file)
software objects, not `project` software objects.  The compilation
database is used in conjunction with projects; conflating the two
is likely dangerous.  I doubt someone would pass a compilation database
parameter when using a single-file software object, but if so, this
is more much more likely to be a mistake  than valid.