Use UIOP/FLEXI-STREAMS for read-/write-shell-file
This makes access to the separate implementations more uniform and it
also (likely) provides coverage of more implementation than just ccl,
sbcl, and ecl.

This also makes it possible to transparently stream bytes or
characters to/from shell pipes on CCL (previously this was only true
for SBCL).

Finally this adds a new `write-shell' function to sel/utility, and it
unifies the implementation of all of the {read,write}-shell{,-file}
functions behind a single new internal `io-shell' function.