Minor update to LD_LIBRARY_PATH environment var
Do not include a trailing colon when setting the LD_LIBRARY_PATH
environment variable in the Dockerfile.  If the LD_LIBRARY_PATH is
empty, we would previously leave a hanging colon, which, within the
Docker container, adds the current working directory to the
LD_LIBRARY_PATH.  This causes issues when running Bug-Injector with
projects which create shared objects (e.g. zlib) as the instrumented
shared libraries are placed on the LD_LIBRARY_PATH.