Coq/serapi-io: reduce serapi startup time
The two-line prologue that's printed on SerAPI startup doesn't end with
a "terminating" response string, so was always taking the full
serapi-timeout to startup. Using read-with-timeout means that we will
read those two lines almost immediately after they're available
(note that if for some reason they are absent, we will end up taking
2 * serapi-timeout to start up instead, but this doesn't seem to happen
in practice).