Fix style warnings
The DEFCLASS form from DEFINE-MUTATION was not showing up
as a top level form.  Make it one so the class is visible
when following forms are compiled.

Also, explicitly provide DEFGENERIC for PICKER and TARGETER,
which are not yet implicitly defined at this point when no
mutations have been defined.  It could be worthwhile to stick
these slots in MUTATION itself.