Add num-traces parameter to trace collection
1. Allow the client to specify a :num-traces parameter to the
trace collection routines.  This allows the client to control
the total number of traces collected independently of how many
test cases are executed.

2. Rename :max to :max-trace-points.  This more clearly identifies
this parameter controls the number of trace points collected per trace
rather than some other max (i.e. maximum number of traces).

3. Remove the :num-threads parameter from `collect-traces`.  This
parameter is unused and is does not work properly as there may be
race conditions in the host program test suite and the trace database.
Further, after adding a :num-traces parameter to the `collect-traces`
routine, we need to know prior to running each test case how many
total traces can be collected; by far, the easiest way to get this
value is to run the test cases one-by-one.

4. Rename `collect-trace` to `collect-test-case-traces`.  I think
this more clearly reflects the reality that we are collecting 1 or
more traces from a single test case.