ASM-SUPER-MUTANT Fitness evaluation improvements.
Add support for linking at different addresses.

Improved initialization of pages:
Read-only and read-write pages are used appropriately.
This handles cases where a page is already being used
read-only, and we don't want to unnecessarily change it to
read-write.

Fitness improvements:
Added support for filtering out dead areas of the stack.
Added save/restore all registers around fitness tests.
Moved the stack location using alloca() to try to avoid stack collision.

Callee-save registers always preserved:
If callee-save registers are not specified in the
i/o spec for testcases, add them and set them to an
arbitrary value on entry, and ensure they retain that
value when the function exits.

Minor changes:
Changed a constant to a special variable for run-time modification.
Commented out an assertion in RUN-FACTORIAL-CL-FUNC which causes the Lisp
process to quit when running ccl. This is because
SHOW-HELP-FOR-FACT-CL-ENTRY calls UIOP:QUIT.