Add a genome-lines-mixin
This mixin adds a slot that is the source file broken
down into lines.  The slot is automatically invalidated
when genome is updated, and (on demand) is obtained
from the genome by decomposing it into a vector of
strings (not including the line terminator).   There
is also a slot for indicating if the last line was
terminated by the line terminator (default Newline)
or by the end of the genome.