Save space reading strings from clang-mutate
By default, string constants are read as (string character),
with the full unicode character set.  But SBCL's reader has
the option to detect when the string is all base chars, and
return a (string base-char) instead.  This is the common
case, and uses less space.  Use a tweaked readtable on sbcl
to do this.  On other lisps, there will be no change.