Check lisp-format in CI pipeline
This adds a new "check-format" stage to the GitLab pipeline which will
fail if lisp-format would change any file modified by the commit.  See
the .ci/Dockerfile.formatter file for the implementation of this step.

This also removes the lisp-format and git-lisp-format scripts
previously committed under the bin/ directory as they are now
downloaded by the Dockerfile mentioned above.