Reduce instrumentation compilation failures
1. Prefix all trace writing data structures/functions with '__' to
reduce probability of name conflicts.

2. Forward declare instrumentation functions/variables at the top of
the file and provide the implementation at the end of the file.  This
reduces the number of header files which need to be included at the
top of the file, which may interfere with the operation of the host
program.  The headers at the top of the file have been minimized to
the absolute minimum required for the forward declarations of the
current instrumentation setup; this necessicates using void* pointers
for instrumentation variables and casting them to the correct
types in the implementation to reduce header file inclusion.

3. Move write_buffer_size function into SEL.  This ensures all trace
writing functions are in one location.

With these changes, no warnings are given with -Wall -Wpedantic using
gcc, g++, and clang++.