Make creating and using store files easier
bin/sel-dump-store is a simple utility for dumping
software as a store file.

Modify create-software to treat files of type .store
as store files, not as source or project files.  Other
sw-specific command line options are ignored when loading
a store file, as they should already be present in the
stored object(s).  We may want to revisit this if it proves
necessary to redefine some or all of those properties, for
example if the store file has been moved to another system.