Move new-clang closer to compatibility with old
Add a new type, nct+, that is intended to be a drop-in
replacement for sel/sw/clang:clang-type.  These objects
include information that is not properly part of the
types, like storage classes.

Add macro table to new-clang objects.  Instead of
getting the macros from clang-mutate, find the #define
statements in the source text.

Fix issue #27 (make :file attribute optional)

Add various methods that were needed.