Make replace-all generic
There are two reasons to do this.  The first is the need
to handle rebinding that calls replace-all to substitute
strings.  In new-clang, the replacements can be other things,
so new methods are needed.
Also, replace-all can be made faster (and use less space)
if all the strings are base strings.