Fix up processing of typedef declations
These are funky in the json.  There can be a type declaration
that is placed before the typedef in the json, but that overlaps
with the typedef's source range.  Fix this by moving that part
of the AST under the typedef.  Also, the elaborated type for
the typedef is in the "inner" array for the typedef, but does not
correspond to any source text.  Add a method to just skip over it.