Fix MULTI-THREADED-CLANG-INSTRUMENT test
The *make-statement-fn* global was not bound properly in threads.
Per the bordeax threads docs:

The interaction between threads and dynamic variables is in some cases
complex, and depends on whether the variable has only a global binding
(as established by e.g. DEFVAR/DEFPARAMETER/top-level SETQ) or has
been bound locally (e.g. with LET or LET*) in the calling thread.

* Global bindings are shared between threads: the initial value of a
global variable in the new thread will be the same as in the parent,
and an assignment to such a variable in any thread will be visible to
all threads in which the global binding is visible.

* Local bindings are local to the thread they are introduced in, except
that

* Local bindings in the the caller of MAKE-THREAD may or may not be
shared with the new thread that it creates: this is
implementation-defined.  Portable code should not depend on particular
behaviour in this case, nor should it assign to such variables without
first rebinding them in the new thread.