Fixes to include path handling
1. new-clang `ast-includes` recursively computes required includes to
match the behavior of old clang.

2. The `include-dirs` field on `new-clang` software has been removed.
There were a multitude of ways this could get out of sync with the
include path compiler flags and this information is a duplicate of
the flags.

3. Include and library search path flags are now always (1) canonicalized,
(2) absolute, and (3) directories.

4. clang default system include search paths are utilized in required
includes calculations.  Previously, ast-includes-in-obj would return
"/usr/include/string.h" when a function in <string.h> was present.