Peer review of ncpair branch
Taking a look at the previous commits on this branch.
I did not look too carefully at type and macro handling
as I suspect any issues with these two items will appear
when testing the scion creator.

This changeset:

1. Attempts to ensure < 80 character lines.
2. Removes unguarded debugging print statements.
3. Removes a few unnecessary let bindings.
4. Removes the `confirm-var-rebinding` function which I believe
is unnecessary and inefficient if we are confident in the
underlying functionality.
5. Fixes a few typos.
6. Removes unnecessary `or` clauses where one part of the clause
is commented out.
7. Improves new-clang flag handling so a special :new-clang-flags
keyword is not required in the tests.
8. Makes a marginal attempt better to group like functions/methods
together (e.g. peel-bananas-tree, find-or-add-type).