Simplification of `ast-includes-in-obj`
Add an `ast-include-for-file` function which returns the
AST file associated within a given AST in a format suitable
for use in a #include directive.  When determining the headers
required for a given AST, find the headers where the AST is
located, where the declaration of the variable associated
with the AST is located, and where the type declaration of
the variable associated with the AST is located.

This commit also removes several no-unused methods/functions.