Remove `source-text` attribute of combined nodes
Leaving this in place would have required additional
special case logic for rebinding variables in addition
to all of the existing special case logic.

Instead, a single special case has been added to `source-text`
while retaining the existing behavior that the children
of combined nodes are ASTs (which was not the case in the
original design).