Remove macros from `clear/update-caches`
When macros are added to a program, the genome is changed
and the entire program will need to be re-parsed.  As
part of the re-parsing, the field will be repopulated.
This field does not need to be cleared when only simple AST
manipulations are performed as they do not add macros
without changing the genome.