Only DeclStmt variable children do not overlap
Do not allow the :Var node children of :DeclStmt ASTs to overlap.
This mimics the old behavior within
clang-mutate/Ast.cpp:update_range_offsets.

For all other :Var ASTs, create a combined node when the source
ranges overlap.  Previously, in the old clang front-end, the
smaller Var would be subsumed as a child of the other(s).

This is required for the proper operation of the `parse-source-snippet`
function in conjunction with the malware scions in Bug-Injector.
Consider the following source snippet:

`int a = 0, b = 1;`

`parse-source-snippet` will now return a single combined node.
Previously, it would return two :var nodes, one with "int a = 0"
and the other with "b = 1".  This led to issue in Bug-Injector
when the child snippets returned by `parse-source-snippet` were
written using a ` (format nil "~{~a;~%~}")` expression; in this
case, "int a = 0;\nb = 1;;" would be printed.  The variable
"b" would be given the default value of 'int' when parsing, and
errors would occur returned when attempting to convert the code
to ASTs if "b" was not originally an 'int'.