Remove ASTs representing __attribute__ specs
Remove ASTs representing __attribute__ specifications from the
AST tree.  These were not represented in "old clang" so I am
comfortable with not representing these in "new clang" to
avoid spending inordinate amounts of time on something
which never worked properly.

This came up in the following context:

```
  1 int f(int a) __attribute__((nothrow));
  2 int g(int b) {
  3     return b;
  4 }
  5
  6 int f(int a) {
  7     return g(a);
  8 }
```

In this case, previously, a giant "Combined" not would be
created as the __attribute__((nothrow)) and it's source
location was associated with both the function declaration
and definition.

See also the openssh nightly test in Bug-Injector which
was failing due to a more complex variation of the
above.