Fix file attribute for locs after #line directives
Fix the file attribute in new-clang-loc structs which
appear after #line directives in the source code.  The
file attribute should remain as the current file
and not the file specified in the line directive.

This commit also updates `remove-non-program-asts` so
that ASTs appearing after a #line directive are kept
around for processing.