Fix Bug-Injector sqlite regression test
In sqlite, there exists an AST related to the vfsList
macro where the :is-macro-expansion-arg is true but
the spelling location appears prior to the macro expansion
location.  This is due to the macro definition being a function
macro itself.  Because the :is-macro-expansion-arg flag is
set, we use the spelling location when creating :Combined
nodes, leading to incorrect :Combined nodes which include
multiple top-level (e.g. function) ASTs.

To resolve this, we need to check if the spelling location
is after the macro expansion location to determine if the
AST is a macro expansion arg.