Fix REPLACE-REMOVES-TRAILING-SEMI-WITH-WHITESPACE
This commit fixes the REPLACE-REMOVES-TRAILING-SEMICOLON-WITH-WHITESPACE
test.  The test was failing due to fallout from 3268d60f.
This commit brings back the is-macro-arg-expansion test which
was removed in 3268d60f while retaining the check that the
spelling location offset comes after the expansion location
offset to allow sqlite to complete successfully.