Removing combined nodes from new-clang
This commit removes combined nodes from new-clang
although they may be added back in the future.
The behavior of new-clang now matches old-clang;
when the source ranges of two ASTs overlap, the smaller
of the two ASTs becomes a child of the other.

This is a punt to allow us to move past the issues
inherent in combined nodes relating to source text
printing.  When two ASTs which overlap are mutated,
it becomes difficult to determine which parts of
each AST should be printed when requesting the
source text of a combined node. We may wish to bring
back combined nodes in the future if a suitable
solution for this is found.  However, for now,
I feel it is appropriate to simply match the
behavior of old clang which seems like a less
time-consuming route.

This commit also removes the
find-var-type-handles-missing-declaration-type
test.  new-clang properly reports an anonymous
type for the AST in question; previously, a
combined node without an AST type was present.