Fix calculation of macro end locations
This commit fixes the computation of macro end locations.
For function-like macros (e.g. #define FOO(X) (X * 5)) the
end location of the macro is computed by finding the end
parenthesis of the macro call.  For instance, in the call
FOO(5), the end location is the final parenthesis.
For non-function-like macros, the end location of the macro
remains the end of the macro name.  For instance, if a macro
"#define mystrcmp strcmp" is in effect and
mystrcmp("foo", "foo") is called, the end location of the
macro is the opening parenthesis.  Previously, this would
be extended to the end paranthesis, which led to issues
when decorating the ASTs with source code.

Further, this commit simplifies macro handling into a single
"fixer" method.