Define server in `define-command-async-rest'
The `define-command-async-rest' macro (previously
`define-command-rest') has long advertised that it defined a command
to actually run the server, but now it actually does.  It does this by
both (1) defining a rest end point using the SEL/REST functionality
and then by (2) calling `define-command-rest-server' to actually
define the function and command-line entry point to launch the server.