Consolidate `original-file` into `original-path`
Include the mixin `file` class as part of `source` software
objects.  The mixin `file` class contains an `original-path`
field which stores the path to the file from which the
software object was populated.  This replaces the
`original-file` field on `source` software objects.

This commit removes the `located-mutate` restart.  I believe
this restart performs a duplicate function to the
*show-update-asts-errors* variable in parseable.lisp.