Split file into two separate mixins
The first mixin, file, is for those software objects where preserving
the original path from which the software object was created is important.

The second mixin, file-w-attributes, is for those software objects
where preserving file attributes (e.g. permissions/modification time)
is important.

The overhead of preserving attributes is non-trivial and not required
for most software objects where only the original path needs to be
preserved.  This provides a 50% performance improvement to auto-merge.