Improve probility of parsing JS files
First attempt to parse as a module.  If this fails,
attempt to parse as a script.  Script mode does not
allow certain keywords (e.g. export) which appear in
modules.  However, parsing as a script does not use
strict mode by default, increasing the probability of
parsing success.  I don't believe there is a good way
to tell a-priori what type of file JS source is,
and the parser will fail quickly if the wrong
parsing mode is utilized.