Removing the old-clang representation
This manifests in the source code as the removal of
"old-clang" methods in clang.lisp, moving new-clang
into clang.lisp, and renaming to replace new-clang
with simply clang.  "new-clang" is now the default
clang representation.

With this commit, portions of clang.lisp were reorganized
to better group like function together.  Further,
this commit removes clang-mutate from the SEL build system
and documentation.  This commit also removes the
much maligned peel/unpeel bananas methods (with a minor
exception localized to an labels binding in clang-w-fodder.lisp),
removes CCL compiler warnings, and adds/improves type
specializations on methods in clang.lisp.

Because old-clang was removed, this commit also removes scripts
related to the regression testing of the new-clang implementation.