Add default values to all defstruct slots
The value of a structure field without an initform is undefined
in standard Common Lisp unless it is initialized in the constructor.
SBCL and CCL initialize them to NIL, but this behavior must not
be depended on, as some lisp (like Clasp) do otherwise.  If
a programmer wants an error to occur if the slot is not initialized
an error initform should be used (such as a call to
ALEXANDRIA:REQUIRED-ARGUMENT, which is intended for just this
use.)