Fix building of documentation
The documentation build fails with the error given below:

```
 #<SB-FORMAT::FMT-CONTROL "~:[Result~;~:*~A~] is a ~/SB-IMPL:PRINT-TYPE/, ~<~%~9T~:;not a ~/SB-IMPL:PRINT-TYPE/.~>"> is not a string designator.
```

The backtrace indicated the build failure was related to the
`fix-compilation` module, and manual delta debugging indicated
that the issue related to the `take` and `take-while` functions
imported from serapeum.

I believe this issue relates to the type specifiers on these
methods within serapeum.  Ultimately, I do not believe the
error message given is a direct consequent; I believe the
message is due to an error in handling the underlying issue.

I'm not sure if this is necessarily the best solution, but
I'm also not sure I'm well suited to make much progress
against this given the opaque nature of the underlying issue.
Paul D. may know more as he opened an SBCL issue for the
error message given [1].  Paul R. may know more as he
is the author of the serapeum functions [2].  It may be
that we don't want to dedicate more time to this which
is why I sent this MR along.

[1] https://bugs.launchpad.net/asdf/+bug/1803728
[2] https://github.com/ruricolist/serapeum/blame/master/sequences.lisp#L989