Removing the `aux-data` field from ASTs
This has been consolidated with the `annotations` field.

There still exists an `ast-attrs` field on clang software objects
which may be further consolidated with `annotations` in the future.
Additionally, clang ASTs are setup differently than LISP/JAVA
ASTs, which utilize the `define-ast` macro.  This may be consolidated
later with the use of functional trees.