Updating parseable/javascript for functional trees
Updating parseable and javascript software objects to utilize
functional trees.  Note that clang software objects do not
utilize functional trees.

This includes the following:

1. Removing the legacy `define-ast` macro and using functional
tree `node` class as the superclass for parseable software
object ASTs.

2. Moving legacy AST mapping operations into clang.lisp (e.g.
`map-ast-with-ancestors`).  We should now utilize fset overrides
of common lisp functions to achieve the same behavior as
these operations.

3. Removing or moving legacy AST -> source range logic.  This
was primarily utilized in BED.

4. Fixing the :referenceddecl field on clang ASTs by ensuring
independent copies of the annotation's field are kept by different
ASTs.

5. Removing the `ast` class specifier from some methods.  Because
clang-ast structs do not inherit from the functional tree base
`ast` class, the type specifier unfortunately can no longer be
utilized.  When clang is updated to use functional trees, we
can re-add these class specifiers (there were far fewer than
I expected).

6. Renaming implementations of tree manipulations in clang.lisp
to utilize the same name and method signature as functional
trees for interoperability (e.g. `replace-ast` -> `with`).

7. Adding implementations of certain methods common to all ASTs
(e.g. `ast-equal-p`, `ast-p`, `ast-hash`) to clang.lisp so clang
and functional tree ASTs can be used interchangeably in most
places.

8. Miscellaneous cleanup such as preferring
`(defgeneric ... (:method ...))` forms.

This is not intended to be an end-all for the conversion to
functional trees and removal of certain legacy tree manipulation
methods.  However, it is intended to be a strong start in this
direction.  Opportunities for future enhancement have been
marked with FIXMEs.