Fix time command in downstream nightlies
The testbot.mk file in the testbot repository
assumes the time command exists at /usr/bin/time
by default.  Install the time package in the SEL
Dockerfile to ensure the command exists in this
location and has the features required by the
testbot Makefile that the built-in shell command
does not.