Allow docs to be built in Ubuntu Docker images
Earlier versions of makeinfo installed by default on Ubuntu
16.04 or 18.04 did not allow menus to be non-optional in texinfo
source files.  An error would occur with the warning

```
warning: node `X' is Y for `Z` but not in menu
```

This commit adds the requisite menus to allow the build to succeed
with the standard version of makeinfo installed by default
on these Ubuntu versions.

Additionally, it fixes the asm super mutant page to use
subsubsection instead of subsection.

This should increase compliance in ensuring the docs are able
to be built by team members using these versions of Ubuntu
as they do not need to test in an arch Docker image.