Fix copying symlinks and nested directories
This commit maintains the functionality of not creating a directory if
there is only one artifact.

Following symlinks is necessary as the original could be changed or
the link could be broken by copying to a new directory.

Keeping the directory structure prevents files with the same name
but different path from being overwritten by each other when they're
copied to BIN.