Make functional tree AST fields read-only
This immutablity of functional tree nodes is required by the
functional trees library.

Note that the `ast-stored-hash` field is a cache based on the
values of immutable fields and is populated in the `ast-hash`
method.  Because the fields upon which this value depends
are immutable, it is safe to set lazily.