Improve performance of `ast-to-list` method
This resolves a performance issue reported by Paul Rodriguez.
The clang `ast-to-list` method utilizes the previous, known
performant solution.  The functional tree AST `ast-to-list`
method utilizes a fold known to be performant as well.

Note that this requires both a functional trees and clang
specialization for now; the plan long term is to make clang
ASTs functional tree nodes to remove the need for a clang-specific
specialization.