Remove `ast-root` from parseable objects
The `genome` field now stores a lazily parsed AST representation
of the code, and this AST representation is returned by `genome`.
`genome-string` is now utilized in those places where a string is
expected for the genome.  Beyond this, various :before/:after/:around
methods for genomes were updated to ensure proper operation.

This changeset also renames `append-to-genome` ->
`append-text-to-genome`, `append-to-genome-preamble` ->
`append-text-to-genome-preamble`, and `prepend-to-genome` ->
`prepend-text-to-genome`.