Utilize functional tree path infrastructure in SEL
Update SEL functional tree ASTs to utilize path infrastructure
within the functional trees library.  This includes updating
`ast-path` to utilize two parameters; the first being the root
of an AST tree and the second being an AST within the tree.
The `ast-path` method for functional trees delegates to the
functional trees `transform-finger` method to find the path.
Fingers are lazily populated when the `ast-path` is requested.
The previous `ast-later-p` method has been renamed `path-later-p`
and takes lists instead of ASTs.

For clang software objects, the existing approach is preserved,
although for compatability purposes all instances of `ast-path`
now take a tree root as an (ignored) first parameter.